requirements.txt
asgiref==3.2.10
Django==3.1.1
django-filter==2.4.0
djangorestframework==3.12.2
importlib-metadata==3.0.0
Markdown==3.3.3
numpy==1.19.2
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.3.1
zipp==3.4.0
